Move StatusHistoryView into it's own file (IOS-2)
